Move fact groups out of Vehicle to separate source files
