+++ src/Vehicle/Vehicle.cc
@@ -745,9 +745,6 @@ void Vehicle::_mavlinkMessageReceived(LinkInterface* link, mavlink_message_t mes
     case MAVLINK_MSG_ID_SCALED_IMU3:
         emit mavlinkScaledImu3(message);
         break;
-    case MAVLINK_MSG_ID_VIBRATION:
-        _handleVibration(message);
-        break;
     case MAVLINK_MSG_ID_EXTENDED_SYS_STATE:
         _handleExtendedSysState(message);
         break;
@@ -756,9 +753,6 @@ void Vehicle::_mavlinkMessageReceived(LinkInterface* link, mavlink_message_t mes
         break;
     case MAVLINK_MSG_ID_COMMAND_LONG:
         _handleCommandLong(message);
-        break;
-    case MAVLINK_MSG_ID_WIND_COV:
-        _handleWindCov(message);
         break;
     case MAVLINK_MSG_ID_LOGGING_DATA:
         _handleMavlinkLoggingData(message);
@@ -778,15 +772,6 @@ void Vehicle::_mavlinkMessageReceived(LinkInterface* link, mavlink_message_t mes
     case MAVLINK_MSG_ID_VFR_HUD:
         _handleVfrHud(message);
         break;
-    case MAVLINK_MSG_ID_SCALED_PRESSURE:
-        _handleScaledPressure(message);
-        break;
-    case MAVLINK_MSG_ID_SCALED_PRESSURE2:
-        _handleScaledPressure2(message);
-        break;
-    case MAVLINK_MSG_ID_SCALED_PRESSURE3:
-        _handleScaledPressure3(message);
-        break;
     case MAVLINK_MSG_ID_CAMERA_IMAGE_CAPTURED:
         _handleCameraImageCaptured(message);
         break;
@@ -804,15 +789,6 @@ void Vehicle::_mavlinkMessageReceived(LinkInterface* link, mavlink_message_t mes
         break;
     case MAVLINK_MSG_ID_ATTITUDE_QUATERNION:
         _handleAttitudeQuaternion(message);
-        break;
-    case MAVLINK_MSG_ID_ATTITUDE_TARGET:
-        _handleAttitudeTarget(message);
-        break;
-    case MAVLINK_MSG_ID_DISTANCE_SENSOR:
-        _handleDistanceSensor(message);
-        break;
-    case MAVLINK_MSG_ID_ESTIMATOR_STATUS:
-        _handleEstimatorStatus(message);
         break;
     case MAVLINK_MSG_ID_STATUSTEXT:
         _handleStatusText(message);
@@ -845,9 +821,6 @@ void Vehicle::_mavlinkMessageReceived(LinkInterface* link, mavlink_message_t mes
 #if !defined(NO_ARDUPILOT_DIALECT)
     case MAVLINK_MSG_ID_CAMERA_FEEDBACK:
         _handleCameraFeedback(message);
-        break;
-    case MAVLINK_MSG_ID_WIND:
-        _handleWind(message);
         break;
 #endif
     }
@@ -1051,98 +1024,6 @@ void Vehicle::_handleVfrHud(mavlink_message_t& message)
     _throttlePctFact.setRawValue(static_cast<int16_t>(vfrHud.throttle));
 }
 
-void Vehicle::_handleEstimatorStatus(mavlink_message_t& message)
-{
-    mavlink_estimator_status_t estimatorStatus;
-    mavlink_msg_estimator_status_decode(&message, &estimatorStatus);
-
-    _estimatorStatusFactGroup.goodAttitudeEstimate()->setRawValue(!!(estimatorStatus.flags & ESTIMATOR_ATTITUDE));
-    _estimatorStatusFactGroup.goodHorizVelEstimate()->setRawValue(!!(estimatorStatus.flags & ESTIMATOR_VELOCITY_HORIZ));
-    _estimatorStatusFactGroup.goodVertVelEstimate()->setRawValue(!!(estimatorStatus.flags & ESTIMATOR_VELOCITY_VERT));
-    _estimatorStatusFactGroup.goodHorizPosRelEstimate()->setRawValue(!!(estimatorStatus.flags & ESTIMATOR_POS_HORIZ_REL));
-    _estimatorStatusFactGroup.goodHorizPosAbsEstimate()->setRawValue(!!(estimatorStatus.flags & ESTIMATOR_POS_HORIZ_ABS));
-    _estimatorStatusFactGroup.goodVertPosAbsEstimate()->setRawValue(!!(estimatorStatus.flags & ESTIMATOR_POS_VERT_ABS));
-    _estimatorStatusFactGroup.goodVertPosAGLEstimate()->setRawValue(!!(estimatorStatus.flags & ESTIMATOR_POS_VERT_AGL));
-    _estimatorStatusFactGroup.goodConstPosModeEstimate()->setRawValue(!!(estimatorStatus.flags & ESTIMATOR_CONST_POS_MODE));
-    _estimatorStatusFactGroup.goodPredHorizPosRelEstimate()->setRawValue(!!(estimatorStatus.flags & ESTIMATOR_PRED_POS_HORIZ_REL));
-    _estimatorStatusFactGroup.goodPredHorizPosAbsEstimate()->setRawValue(!!(estimatorStatus.flags & ESTIMATOR_PRED_POS_HORIZ_ABS));
-    _estimatorStatusFactGroup.gpsGlitch()->setRawValue(estimatorStatus.flags & ESTIMATOR_GPS_GLITCH ? true : false);
-    _estimatorStatusFactGroup.accelError()->setRawValue(!!(estimatorStatus.flags & ESTIMATOR_ACCEL_ERROR));
-    _estimatorStatusFactGroup.velRatio()->setRawValue(estimatorStatus.vel_ratio);
-    _estimatorStatusFactGroup.horizPosRatio()->setRawValue(estimatorStatus.pos_horiz_ratio);
-    _estimatorStatusFactGroup.vertPosRatio()->setRawValue(estimatorStatus.pos_vert_ratio);
-    _estimatorStatusFactGroup.magRatio()->setRawValue(estimatorStatus.mag_ratio);
-    _estimatorStatusFactGroup.haglRatio()->setRawValue(estimatorStatus.hagl_ratio);
-    _estimatorStatusFactGroup.tasRatio()->setRawValue(estimatorStatus.tas_ratio);
-    _estimatorStatusFactGroup.horizPosAccuracy()->setRawValue(estimatorStatus.pos_horiz_accuracy);
-    _estimatorStatusFactGroup.vertPosAccuracy()->setRawValue(estimatorStatus.pos_vert_accuracy);
-
-#if 0
-    typedef enum ESTIMATOR_STATUS_FLAGS
-    {
-        ESTIMATOR_ATTITUDE=1, /* True if the attitude estimate is good | */
-        ESTIMATOR_VELOCITY_HORIZ=2, /* True if the horizontal velocity estimate is good | */
-        ESTIMATOR_VELOCITY_VERT=4, /* True if the  vertical velocity estimate is good | */
-        ESTIMATOR_POS_HORIZ_REL=8, /* True if the horizontal position (relative) estimate is good | */
-        ESTIMATOR_POS_HORIZ_ABS=16, /* True if the horizontal position (absolute) estimate is good | */
-        ESTIMATOR_POS_VERT_ABS=32, /* True if the vertical position (absolute) estimate is good | */
-        ESTIMATOR_POS_VERT_AGL=64, /* True if the vertical position (above ground) estimate is good | */
-        ESTIMATOR_CONST_POS_MODE=128, /* True if the EKF is in a constant position mode and is not using external measurements (eg GPS or optical flow) | */
-        ESTIMATOR_PRED_POS_HORIZ_REL=256, /* True if the EKF has sufficient data to enter a mode that will provide a (relative) position estimate | */
-        ESTIMATOR_PRED_POS_HORIZ_ABS=512, /* True if the EKF has sufficient data to enter a mode that will provide a (absolute) position estimate | */
-        ESTIMATOR_GPS_GLITCH=1024, /* True if the EKF has detected a GPS glitch | */
-        ESTIMATOR_ACCEL_ERROR=2048, /* True if the EKF has detected bad accelerometer data | */
-
-        typedef struct __mavlink_estimator_status_t {
-            uint64_t time_usec; /*< Timestamp (micros since boot or Unix epoch)*/
-            float vel_ratio; /*< Velocity innovation test ratio*/
-            float pos_horiz_ratio; /*< Horizontal position innovation test ratio*/
-            float pos_vert_ratio; /*< Vertical position innovation test ratio*/
-            float mag_ratio; /*< Magnetometer innovation test ratio*/
-            float hagl_ratio; /*< Height above terrain innovation test ratio*/
-            float tas_ratio; /*< True airspeed innovation test ratio*/
-            float pos_horiz_accuracy; /*< Horizontal position 1-STD accuracy relative to the EKF local origin (m)*/
-            float pos_vert_accuracy; /*< Vertical position 1-STD accuracy relative to the EKF local origin (m)*/
-            uint16_t flags; /*< Integer bitmask indicating which EKF outputs are valid. See definition for ESTIMATOR_STATUS_FLAGS.*/
-        } mavlink_estimator_status_t;
-    };
-#endif
-}
-
-void Vehicle::_handleDistanceSensor(mavlink_message_t& message)
-{
-    mavlink_distance_sensor_t distanceSensor;
-
-    mavlink_msg_distance_sensor_decode(&message, &distanceSensor);
-
-    struct orientation2Fact_s {
-        MAV_SENSOR_ORIENTATION  orientation;
-        Fact*                   fact;
-    };
-
-    orientation2Fact_s rgOrientation2Fact[] =
-    {
-        { MAV_SENSOR_ROTATION_NONE,         _distanceSensorFactGroup.rotationNone() },
-        { MAV_SENSOR_ROTATION_YAW_45,       _distanceSensorFactGroup.rotationYaw45() },
-        { MAV_SENSOR_ROTATION_YAW_90,       _distanceSensorFactGroup.rotationYaw90() },
-        { MAV_SENSOR_ROTATION_YAW_135,      _distanceSensorFactGroup.rotationYaw135() },
-        { MAV_SENSOR_ROTATION_YAW_180,      _distanceSensorFactGroup.rotationYaw180() },
-        { MAV_SENSOR_ROTATION_YAW_225,      _distanceSensorFactGroup.rotationYaw225() },
-        { MAV_SENSOR_ROTATION_YAW_270,      _distanceSensorFactGroup.rotationYaw270() },
-        { MAV_SENSOR_ROTATION_YAW_315,      _distanceSensorFactGroup.rotationYaw315() },
-        { MAV_SENSOR_ROTATION_PITCH_90,     _distanceSensorFactGroup.rotationPitch90() },
-        { MAV_SENSOR_ROTATION_PITCH_270,    _distanceSensorFactGroup.rotationPitch270() },
-    };
-
-    for (size_t i=0; i<sizeof(rgOrientation2Fact)/sizeof(rgOrientation2Fact[0]); i++) {
-        const orientation2Fact_s& orientation2Fact = rgOrientation2Fact[i];
-        if (orientation2Fact.orientation == distanceSensor.orientation) {
-            orientation2Fact.fact->setRawValue(distanceSensor.current_distance / 100.0); // cm to meters
-        }
-    _distanceSensorFactGroup.maxDistance()->setRawValue(distanceSensor.max_distance/100.0);
-    }
-}
-
 // Ignore warnings from mavlink headers for both GCC/Clang and MSVC
 #ifdef __GNUC__
 
@@ -1160,24 +1041,6 @@ void Vehicle::_handleDistanceSensor(mavlink_message_t& message)
 #else
 #pragma warning(push, 0)
 #endif
-
-void Vehicle::_handleAttitudeTarget(mavlink_message_t& message)
-{
-    mavlink_attitude_target_t attitudeTarget;
-
-    mavlink_msg_attitude_target_decode(&message, &attitudeTarget);
-
-    float roll, pitch, yaw;
-    mavlink_quaternion_to_euler(attitudeTarget.q, &roll, &pitch, &yaw);
-
-    _setpointFactGroup.roll()->setRawValue(qRadiansToDegrees(roll));
-    _setpointFactGroup.pitch()->setRawValue(qRadiansToDegrees(pitch));
-    _setpointFactGroup.yaw()->setRawValue(qRadiansToDegrees(yaw));
-
-    _setpointFactGroup.rollRate()->setRawValue(qRadiansToDegrees(attitudeTarget.body_roll_rate));
-    _setpointFactGroup.pitchRate()->setRawValue(qRadiansToDegrees(attitudeTarget.body_pitch_rate));
-    _setpointFactGroup.yawRate()->setRawValue(qRadiansToDegrees(attitudeTarget.body_yaw_rate));
-}
 
 void Vehicle::_handleAttitudeWorker(double rollRadians, double pitchRadians, double yawRadians)
 {
@@ -1261,15 +1124,6 @@ void Vehicle::_handleGpsRawInt(mavlink_message_t& message)
             }
         }
     }
-
-    _gpsFactGroup.lat()->setRawValue(gpsRawInt.lat * 1e-7);
-    _gpsFactGroup.lon()->setRawValue(gpsRawInt.lon * 1e-7);
-    _gpsFactGroup.mgrs()->setRawValue(convertGeoToMGRS(QGeoCoordinate(gpsRawInt.lat * 1e-7, gpsRawInt.lon * 1e-7)));
-    _gpsFactGroup.count()->setRawValue(gpsRawInt.satellites_visible == 255 ? 0 : gpsRawInt.satellites_visible);
-    _gpsFactGroup.hdop()->setRawValue(gpsRawInt.eph == UINT16_MAX ? qQNaN() : gpsRawInt.eph / 100.0);
-    _gpsFactGroup.vdop()->setRawValue(gpsRawInt.epv == UINT16_MAX ? qQNaN() : gpsRawInt.epv / 100.0);
-    _gpsFactGroup.courseOverGround()->setRawValue(gpsRawInt.cog == UINT16_MAX ? qQNaN() : gpsRawInt.cog / 100.0);
-    _gpsFactGroup.lock()->setRawValue(gpsRawInt.fix_type);
 }
 
 void Vehicle::_handleGlobalPositionInt(mavlink_message_t& message)
@@ -1340,15 +1194,6 @@ void Vehicle::_handleHighLatency(mavlink_message_t& message)
     _headingFact.setRawValue((double)highLatency.heading * 2.0);
     _altitudeRelativeFact.setRawValue(qQNaN());
     _altitudeAMSLFact.setRawValue(coordinate.altitude);
-
-    _windFactGroup.speed()->setRawValue((double)highLatency.airspeed / 5.0);
-
-    _temperatureFactGroup.temperature1()->setRawValue(highLatency.temperature_air);
-
-    _gpsFactGroup.lat()->setRawValue(coordinate.latitude);
-    _gpsFactGroup.lon()->setRawValue(coordinate.longitude);
-    _gpsFactGroup.mgrs()->setRawValue(convertGeoToMGRS(QGeoCoordinate(coordinate.latitude, coordinate.longitude)));
-    _gpsFactGroup.count()->setRawValue(0);
 }
 
 void Vehicle::_handleHighLatency2(mavlink_message_t& message)
@@ -1385,18 +1230,6 @@ void Vehicle::_handleHighLatency2(mavlink_message_t& message)
     _headingFact.setRawValue((double)highLatency2.heading * 2.0);
     _altitudeRelativeFact.setRawValue(qQNaN());
     _altitudeAMSLFact.setRawValue(highLatency2.altitude);
-
-    _windFactGroup.direction()->setRawValue((double)highLatency2.wind_heading * 2.0);
-    _windFactGroup.speed()->setRawValue((double)highLatency2.windspeed / 5.0);
-
-    _temperatureFactGroup.temperature1()->setRawValue(highLatency2.temperature_air);
-
-    _gpsFactGroup.lat()->setRawValue(highLatency2.latitude * 1e-7);
-    _gpsFactGroup.lon()->setRawValue(highLatency2.longitude * 1e-7);
-    _gpsFactGroup.mgrs()->setRawValue(convertGeoToMGRS(QGeoCoordinate(highLatency2.latitude * 1e-7, highLatency2.longitude * 1e-7)));
-    _gpsFactGroup.count()->setRawValue(0);
-    _gpsFactGroup.hdop()->setRawValue(highLatency2.eph == UINT8_MAX ? qQNaN() : highLatency2.eph / 10.0);
-    _gpsFactGroup.vdop()->setRawValue(highLatency2.epv == UINT8_MAX ? qQNaN() : highLatency2.epv / 10.0);
 
     struct failure2Sensor_s {
         HL_FAILURE_FLAG         failureBit;
@@ -1558,53 +1391,6 @@ void Vehicle::_handleExtendedSysState(mavlink_message_t& message)
         }
     }
 }
-
-void Vehicle::_handleVibration(mavlink_message_t& message)
-{
-    mavlink_vibration_t vibration;
-    mavlink_msg_vibration_decode(&message, &vibration);
-
-    _vibrationFactGroup.xAxis()->setRawValue(vibration.vibration_x);
-    _vibrationFactGroup.yAxis()->setRawValue(vibration.vibration_y);
-    _vibrationFactGroup.zAxis()->setRawValue(vibration.vibration_z);
-    _vibrationFactGroup.clipCount1()->setRawValue(vibration.clipping_0);
-    _vibrationFactGroup.clipCount2()->setRawValue(vibration.clipping_1);
-    _vibrationFactGroup.clipCount3()->setRawValue(vibration.clipping_2);
-}
-
-void Vehicle::_handleWindCov(mavlink_message_t& message)
-{
-    mavlink_wind_cov_t wind;
-    mavlink_msg_wind_cov_decode(&message, &wind);
-
-    float direction = qRadiansToDegrees(qAtan2(wind.wind_y, wind.wind_x));
-    float speed = qSqrt(qPow(wind.wind_x, 2) + qPow(wind.wind_y, 2));
-
-    if (direction < 0) {
-        direction += 360;
-    }
-
-    _windFactGroup.direction()->setRawValue(direction);
-    _windFactGroup.speed()->setRawValue(speed);
-    _windFactGroup.verticalSpeed()->setRawValue(0);
-}
-
-#if !defined(NO_ARDUPILOT_DIALECT)
-void Vehicle::_handleWind(mavlink_message_t& message)
-{
-    mavlink_wind_t wind;
-    mavlink_msg_wind_decode(&message, &wind);
-
-    // We don't want negative wind angles
-    float direction = wind.direction;
-    if (direction < 0) {
-        direction += 360;
-    }
-    _windFactGroup.direction()->setRawValue(direction);
-    _windFactGroup.speed()->setRawValue(wind.speed);
-    _windFactGroup.verticalSpeed()->setRawValue(wind.speed_z);
-}
-#endif
 
 bool Vehicle::_apmArmingNotRequired()
 {
@@ -1940,24 +1726,6 @@ void Vehicle::_handleRCChannels(mavlink_message_t& message)
 #else
 #pragma warning(pop, 0)
 #endif
-
-void Vehicle::_handleScaledPressure(mavlink_message_t& message) {
-    mavlink_scaled_pressure_t pressure;
-    mavlink_msg_scaled_pressure_decode(&message, &pressure);
-    _temperatureFactGroup.temperature1()->setRawValue(pressure.temperature / 100.0);
-}
-
-void Vehicle::_handleScaledPressure2(mavlink_message_t& message) {
-    mavlink_scaled_pressure2_t pressure;
-    mavlink_msg_scaled_pressure2_decode(&message, &pressure);
-    _temperatureFactGroup.temperature2()->setRawValue(pressure.temperature / 100.0);
-}
-
-void Vehicle::_handleScaledPressure3(mavlink_message_t& message) {
-    mavlink_scaled_pressure3_t pressure;
-    mavlink_msg_scaled_pressure3_decode(&message, &pressure);
-    _temperatureFactGroup.temperature3()->setRawValue(pressure.temperature / 100.0);
-}
 
 bool Vehicle::_containsLink(LinkInterface* link)
 {
@@ -3703,43 +3471,6 @@ void Vehicle::setVtolInFwdFlight(bool vtolInFwdFlight)
     }
 }
 
-const char* VehicleGPSFactGroup::_latFactName =                 "lat";
-const char* VehicleGPSFactGroup::_lonFactName =                 "lon";
-const char* VehicleGPSFactGroup::_mgrsFactName =                "mgrs";
-const char* VehicleGPSFactGroup::_hdopFactName =                "hdop";
-const char* VehicleGPSFactGroup::_vdopFactName =                "vdop";
-const char* VehicleGPSFactGroup::_courseOverGroundFactName =    "courseOverGround";
-const char* VehicleGPSFactGroup::_countFactName =               "count";
-const char* VehicleGPSFactGroup::_lockFactName =                "lock";
-
-VehicleGPSFactGroup::VehicleGPSFactGroup(QObject* parent)
-    : FactGroup(1000, ":/json/Vehicle/GPSFact.json", parent)
-    , _latFact              (0, _latFactName,               FactMetaData::valueTypeDouble)
-    , _lonFact              (0, _lonFactName,               FactMetaData::valueTypeDouble)
-    , _mgrsFact             (0, _mgrsFactName,              FactMetaData::valueTypeString)
-    , _hdopFact             (0, _hdopFactName,              FactMetaData::valueTypeDouble)
-    , _vdopFact             (0, _vdopFactName,              FactMetaData::valueTypeDouble)
-    , _courseOverGroundFact (0, _courseOverGroundFactName,  FactMetaData::valueTypeDouble)
-    , _countFact            (0, _countFactName,             FactMetaData::valueTypeInt32)
-    , _lockFact             (0, _lockFactName,              FactMetaData::valueTypeInt32)
-{
-    _addFact(&_latFact,                 _latFactName);
-    _addFact(&_lonFact,                 _lonFactName);
-    _addFact(&_mgrsFact,                _mgrsFactName);
-    _addFact(&_hdopFact,                _hdopFactName);
-    _addFact(&_vdopFact,                _vdopFactName);
-    _addFact(&_courseOverGroundFact,    _courseOverGroundFactName);
-    _addFact(&_lockFact,                _lockFactName);
-    _addFact(&_countFact,               _countFactName);
-
-    _latFact.setRawValue(std::numeric_limits<float>::quiet_NaN());
-    _lonFact.setRawValue(std::numeric_limits<float>::quiet_NaN());
-    _mgrsFact.setRawValue("");
-    _hdopFact.setRawValue(std::numeric_limits<float>::quiet_NaN());
-    _vdopFact.setRawValue(std::numeric_limits<float>::quiet_NaN());
-    _courseOverGroundFact.setRawValue(std::numeric_limits<float>::quiet_NaN());
-}
-
 void Vehicle::startMavlinkLog()
 {
     sendMavCommand(_defaultComponentId, MAV_CMD_LOGGING_START, false /* showError */);
@@ -4378,246 +4109,4 @@ void Vehicle::sendJoystickDataThreadSafe(float roll, float pitch, float yaw, flo
                 static_cast<int16_t>(newYawCommand),
                 buttons);
     sendMessageOnLinkThreadSafe(priorityLink(), message);
-}
-
-const char* VehicleWindFactGroup::_directionFactName =      "direction";
-const char* VehicleWindFactGroup::_speedFactName =          "speed";
-const char* VehicleWindFactGroup::_verticalSpeedFactName =  "verticalSpeed";
-
-VehicleWindFactGroup::VehicleWindFactGroup(QObject* parent)
-    : FactGroup(1000, ":/json/Vehicle/WindFact.json", parent)
-    , _directionFact    (0, _directionFactName,     FactMetaData::valueTypeDouble)
-    , _speedFact        (0, _speedFactName,         FactMetaData::valueTypeDouble)
-    , _verticalSpeedFact(0, _verticalSpeedFactName, FactMetaData::valueTypeDouble)
-{
-    _addFact(&_directionFact,       _directionFactName);
-    _addFact(&_speedFact,           _speedFactName);
-    _addFact(&_verticalSpeedFact,   _verticalSpeedFactName);
-
-    // Start out as not available "--.--"
-    _directionFact.setRawValue      (qQNaN());
-    _speedFact.setRawValue          (qQNaN());
-    _verticalSpeedFact.setRawValue  (qQNaN());
-}
-
-const char* VehicleVibrationFactGroup::_xAxisFactName =      "xAxis";
-const char* VehicleVibrationFactGroup::_yAxisFactName =      "yAxis";
-const char* VehicleVibrationFactGroup::_zAxisFactName =      "zAxis";
-const char* VehicleVibrationFactGroup::_clipCount1FactName = "clipCount1";
-const char* VehicleVibrationFactGroup::_clipCount2FactName = "clipCount2";
-const char* VehicleVibrationFactGroup::_clipCount3FactName = "clipCount3";
-
-VehicleVibrationFactGroup::VehicleVibrationFactGroup(QObject* parent)
-    : FactGroup(1000, ":/json/Vehicle/VibrationFact.json", parent)
-    , _xAxisFact        (0, _xAxisFactName,         FactMetaData::valueTypeDouble)
-    , _yAxisFact        (0, _yAxisFactName,         FactMetaData::valueTypeDouble)
-    , _zAxisFact        (0, _zAxisFactName,         FactMetaData::valueTypeDouble)
-    , _clipCount1Fact   (0, _clipCount1FactName,    FactMetaData::valueTypeUint32)
-    , _clipCount2Fact   (0, _clipCount2FactName,    FactMetaData::valueTypeUint32)
-    , _clipCount3Fact   (0, _clipCount3FactName,    FactMetaData::valueTypeUint32)
-{
-    _addFact(&_xAxisFact,       _xAxisFactName);
-    _addFact(&_yAxisFact,       _yAxisFactName);
-    _addFact(&_zAxisFact,       _zAxisFactName);
-    _addFact(&_clipCount1Fact,  _clipCount1FactName);
-    _addFact(&_clipCount2Fact,  _clipCount2FactName);
-    _addFact(&_clipCount3Fact,  _clipCount3FactName);
-
-    // Start out as not available "--.--"
-    _xAxisFact.setRawValue(qQNaN());
-    _yAxisFact.setRawValue(qQNaN());
-    _zAxisFact.setRawValue(qQNaN());
-}
-
-const char* VehicleTemperatureFactGroup::_temperature1FactName =      "temperature1";
-const char* VehicleTemperatureFactGroup::_temperature2FactName =      "temperature2";
-const char* VehicleTemperatureFactGroup::_temperature3FactName =      "temperature3";
-
-VehicleTemperatureFactGroup::VehicleTemperatureFactGroup(QObject* parent)
-    : FactGroup(1000, ":/json/Vehicle/TemperatureFact.json", parent)
-    , _temperature1Fact    (0, _temperature1FactName,     FactMetaData::valueTypeDouble)
-    , _temperature2Fact    (0, _temperature2FactName,     FactMetaData::valueTypeDouble)
-    , _temperature3Fact    (0, _temperature3FactName,     FactMetaData::valueTypeDouble)
-{
-    _addFact(&_temperature1Fact,       _temperature1FactName);
-    _addFact(&_temperature2Fact,       _temperature2FactName);
-    _addFact(&_temperature3Fact,       _temperature3FactName);
-
-    // Start out as not available "--.--"
-    _temperature1Fact.setRawValue      (qQNaN());
-    _temperature2Fact.setRawValue      (qQNaN());
-    _temperature3Fact.setRawValue      (qQNaN());
-}
-
-const char* VehicleClockFactGroup::_currentTimeFactName = "currentTime";
-const char* VehicleClockFactGroup::_currentDateFactName = "currentDate";
-
-VehicleClockFactGroup::VehicleClockFactGroup(QObject* parent)
-    : FactGroup(1000, ":/json/Vehicle/ClockFact.json", parent)
-    , _currentTimeFact  (0, _currentTimeFactName,    FactMetaData::valueTypeString)
-    , _currentDateFact  (0, _currentDateFactName,    FactMetaData::valueTypeString)
-{
-    _addFact(&_currentTimeFact, _currentTimeFactName);
-    _addFact(&_currentDateFact, _currentDateFactName);
-
-    // Start out as not available "--.--"
-    _currentTimeFact.setRawValue    (std::numeric_limits<float>::quiet_NaN());
-    _currentDateFact.setRawValue    (std::numeric_limits<float>::quiet_NaN());
-}
-
-void VehicleClockFactGroup::_updateAllValues()
-{
-    _currentTimeFact.setRawValue(QTime::currentTime().toString());
-    _currentDateFact.setRawValue(QDateTime::currentDateTime().toString(QLocale::system().dateFormat(QLocale::ShortFormat)));
-
-    FactGroup::_updateAllValues();
-}
-
-const char* VehicleSetpointFactGroup::_rollFactName =       "roll";
-const char* VehicleSetpointFactGroup::_pitchFactName =      "pitch";
-const char* VehicleSetpointFactGroup::_yawFactName =        "yaw";
-const char* VehicleSetpointFactGroup::_rollRateFactName =   "rollRate";
-const char* VehicleSetpointFactGroup::_pitchRateFactName =  "pitchRate";
-const char* VehicleSetpointFactGroup::_yawRateFactName =    "yawRate";
-
-VehicleSetpointFactGroup::VehicleSetpointFactGroup(QObject* parent)
-    : FactGroup     (1000, ":/json/Vehicle/SetpointFact.json", parent)
-    , _rollFact     (0, _rollFactName,      FactMetaData::valueTypeDouble)
-    , _pitchFact    (0, _pitchFactName,     FactMetaData::valueTypeDouble)
-    , _yawFact      (0, _yawFactName,       FactMetaData::valueTypeDouble)
-    , _rollRateFact (0, _rollRateFactName,  FactMetaData::valueTypeDouble)
-    , _pitchRateFact(0, _pitchRateFactName, FactMetaData::valueTypeDouble)
-    , _yawRateFact  (0, _yawRateFactName,   FactMetaData::valueTypeDouble)
-{
-    _addFact(&_rollFact,        _rollFactName);
-    _addFact(&_pitchFact,       _pitchFactName);
-    _addFact(&_yawFact,         _yawFactName);
-    _addFact(&_rollRateFact,    _rollRateFactName);
-    _addFact(&_pitchRateFact,   _pitchRateFactName);
-    _addFact(&_yawRateFact,     _yawRateFactName);
-
-    // Start out as not available "--.--"
-    _rollFact.setRawValue(qQNaN());
-    _pitchFact.setRawValue(qQNaN());
-    _yawFact.setRawValue(qQNaN());
-    _rollRateFact.setRawValue(qQNaN());
-    _pitchRateFact.setRawValue(qQNaN());
-    _yawRateFact.setRawValue(qQNaN());
-}
-
-const char* VehicleDistanceSensorFactGroup::_rotationNoneFactName =     "rotationNone";
-const char* VehicleDistanceSensorFactGroup::_rotationYaw45FactName =    "rotationYaw45";
-const char* VehicleDistanceSensorFactGroup::_rotationYaw90FactName =    "rotationYaw90";
-const char* VehicleDistanceSensorFactGroup::_rotationYaw135FactName =   "rotationYaw135";
-const char* VehicleDistanceSensorFactGroup::_rotationYaw180FactName =   "rotationYaw180";
-const char* VehicleDistanceSensorFactGroup::_rotationYaw225FactName =   "rotationYaw225";
-const char* VehicleDistanceSensorFactGroup::_rotationYaw270FactName =   "rotationYaw270";
-const char* VehicleDistanceSensorFactGroup::_rotationYaw315FactName =   "rotationYaw315";
-const char* VehicleDistanceSensorFactGroup::_rotationPitch90FactName =  "rotationPitch90";
-const char* VehicleDistanceSensorFactGroup::_rotationPitch270FactName = "rotationPitch270";
-const char* VehicleDistanceSensorFactGroup::_maxDistanceFactName =      "maxDistance";
-
-VehicleDistanceSensorFactGroup::VehicleDistanceSensorFactGroup(QObject* parent)
-    : FactGroup             (1000, ":/json/Vehicle/DistanceSensorFact.json", parent)
-    , _rotationNoneFact     (0, _rotationNoneFactName,      FactMetaData::valueTypeDouble)
-    , _rotationYaw45Fact    (0, _rotationYaw45FactName,     FactMetaData::valueTypeDouble)
-    , _rotationYaw90Fact    (0, _rotationYaw90FactName,     FactMetaData::valueTypeDouble)
-    , _rotationYaw135Fact   (0, _rotationYaw135FactName,    FactMetaData::valueTypeDouble)
-    , _rotationYaw180Fact   (0, _rotationYaw180FactName,    FactMetaData::valueTypeDouble)
-    , _rotationYaw225Fact   (0, _rotationYaw225FactName,    FactMetaData::valueTypeDouble)
-    , _rotationYaw270Fact   (0, _rotationYaw270FactName,    FactMetaData::valueTypeDouble)
-    , _rotationYaw315Fact   (0, _rotationYaw315FactName,    FactMetaData::valueTypeDouble)
-    , _rotationPitch90Fact  (0, _rotationPitch90FactName,   FactMetaData::valueTypeDouble)
-    , _rotationPitch270Fact (0, _rotationPitch270FactName,  FactMetaData::valueTypeDouble)
-    , _maxDistanceFact      (0, _maxDistanceFactName,       FactMetaData::valueTypeDouble)
-{
-    _addFact(&_rotationNoneFact,        _rotationNoneFactName);
-    _addFact(&_rotationYaw45Fact,       _rotationYaw45FactName);
-    _addFact(&_rotationYaw90Fact,       _rotationYaw90FactName);
-    _addFact(&_rotationYaw135Fact,      _rotationYaw135FactName);
-    _addFact(&_rotationYaw180Fact,      _rotationYaw180FactName);
-    _addFact(&_rotationYaw225Fact,      _rotationYaw225FactName);
-    _addFact(&_rotationYaw270Fact,      _rotationYaw270FactName);
-    _addFact(&_rotationYaw315Fact,      _rotationYaw315FactName);
-    _addFact(&_rotationPitch90Fact,     _rotationPitch90FactName);
-    _addFact(&_rotationPitch270Fact,    _rotationPitch270FactName);
-    _addFact(&_maxDistanceFact,         _maxDistanceFactName);
-
-    // Start out as not available "--.--"
-    _rotationNoneFact.setRawValue(qQNaN());
-    _rotationYaw45Fact.setRawValue(qQNaN());
-    _rotationYaw135Fact.setRawValue(qQNaN());
-    _rotationYaw90Fact.setRawValue(qQNaN());
-    _rotationYaw180Fact.setRawValue(qQNaN());
-    _rotationYaw225Fact.setRawValue(qQNaN());
-    _rotationYaw270Fact.setRawValue(qQNaN());
-    _rotationPitch90Fact.setRawValue(qQNaN());
-    _rotationPitch270Fact.setRawValue(qQNaN());
-    _maxDistanceFact.setRawValue(qQNaN());
-}
-
-const char* VehicleEstimatorStatusFactGroup::_goodAttitudeEstimateFactName =        "goodAttitudeEsimate";
-const char* VehicleEstimatorStatusFactGroup::_goodHorizVelEstimateFactName =        "goodHorizVelEstimate";
-const char* VehicleEstimatorStatusFactGroup::_goodVertVelEstimateFactName =         "goodVertVelEstimate";
-const char* VehicleEstimatorStatusFactGroup::_goodHorizPosRelEstimateFactName =     "goodHorizPosRelEstimate";
-const char* VehicleEstimatorStatusFactGroup::_goodHorizPosAbsEstimateFactName =     "goodHorizPosAbsEstimate";
-const char* VehicleEstimatorStatusFactGroup::_goodVertPosAbsEstimateFactName =      "goodVertPosAbsEstimate";
-const char* VehicleEstimatorStatusFactGroup::_goodVertPosAGLEstimateFactName =      "goodVertPosAGLEstimate";
-const char* VehicleEstimatorStatusFactGroup::_goodConstPosModeEstimateFactName =    "goodConstPosModeEstimate";
-const char* VehicleEstimatorStatusFactGroup::_goodPredHorizPosRelEstimateFactName = "goodPredHorizPosRelEstimate";
-const char* VehicleEstimatorStatusFactGroup::_goodPredHorizPosAbsEstimateFactName = "goodPredHorizPosAbsEstimate";
-const char* VehicleEstimatorStatusFactGroup::_gpsGlitchFactName =                   "gpsGlitch";
-const char* VehicleEstimatorStatusFactGroup::_accelErrorFactName =                  "accelError";
-const char* VehicleEstimatorStatusFactGroup::_velRatioFactName =                    "velRatio";
-const char* VehicleEstimatorStatusFactGroup::_horizPosRatioFactName =               "horizPosRatio";
-const char* VehicleEstimatorStatusFactGroup::_vertPosRatioFactName =                "vertPosRatio";
-const char* VehicleEstimatorStatusFactGroup::_magRatioFactName =                    "magRatio";
-const char* VehicleEstimatorStatusFactGroup::_haglRatioFactName =                   "haglRatio";
-const char* VehicleEstimatorStatusFactGroup::_tasRatioFactName =                    "tasRatio";
-const char* VehicleEstimatorStatusFactGroup::_horizPosAccuracyFactName =            "horizPosAccuracy";
-const char* VehicleEstimatorStatusFactGroup::_vertPosAccuracyFactName =             "vertPosAccuracy";
-
-VehicleEstimatorStatusFactGroup::VehicleEstimatorStatusFactGroup(QObject* parent)
-    : FactGroup                         (500, ":/json/Vehicle/EstimatorStatusFactGroup.json", parent)
-    , _goodAttitudeEstimateFact         (0, _goodAttitudeEstimateFactName,          FactMetaData::valueTypeBool)
-    , _goodHorizVelEstimateFact         (0, _goodHorizVelEstimateFactName,          FactMetaData::valueTypeBool)
-    , _goodVertVelEstimateFact          (0, _goodVertVelEstimateFactName,           FactMetaData::valueTypeBool)
-    , _goodHorizPosRelEstimateFact      (0, _goodHorizPosRelEstimateFactName,       FactMetaData::valueTypeBool)
-    , _goodHorizPosAbsEstimateFact      (0, _goodHorizPosAbsEstimateFactName,       FactMetaData::valueTypeBool)
-    , _goodVertPosAbsEstimateFact       (0, _goodVertPosAbsEstimateFactName,        FactMetaData::valueTypeBool)
-    , _goodVertPosAGLEstimateFact       (0, _goodVertPosAGLEstimateFactName,        FactMetaData::valueTypeBool)
-    , _goodConstPosModeEstimateFact     (0, _goodConstPosModeEstimateFactName,      FactMetaData::valueTypeBool)
-    , _goodPredHorizPosRelEstimateFact  (0, _goodPredHorizPosRelEstimateFactName,   FactMetaData::valueTypeBool)
-    , _goodPredHorizPosAbsEstimateFact  (0, _goodPredHorizPosAbsEstimateFactName,   FactMetaData::valueTypeBool)
-    , _gpsGlitchFact                    (0, _gpsGlitchFactName,                     FactMetaData::valueTypeBool)
-    , _accelErrorFact                   (0, _accelErrorFactName,                    FactMetaData::valueTypeBool)
-    , _velRatioFact                     (0, _velRatioFactName,                      FactMetaData::valueTypeFloat)
-    , _horizPosRatioFact                (0, _horizPosRatioFactName,                 FactMetaData::valueTypeFloat)
-    , _vertPosRatioFact                 (0, _vertPosRatioFactName,                  FactMetaData::valueTypeFloat)
-    , _magRatioFact                     (0, _magRatioFactName,                      FactMetaData::valueTypeFloat)
-    , _haglRatioFact                    (0, _haglRatioFactName,                     FactMetaData::valueTypeFloat)
-    , _tasRatioFact                     (0, _tasRatioFactName,                      FactMetaData::valueTypeFloat)
-    , _horizPosAccuracyFact             (0, _horizPosAccuracyFactName,              FactMetaData::valueTypeFloat)
-    , _vertPosAccuracyFact              (0, _vertPosAccuracyFactName,               FactMetaData::valueTypeFloat)
-{
-    _addFact(&_goodAttitudeEstimateFact,        _goodAttitudeEstimateFactName);
-    _addFact(&_goodHorizVelEstimateFact,        _goodHorizVelEstimateFactName);
-    _addFact(&_goodVertVelEstimateFact,         _goodVertVelEstimateFactName);
-    _addFact(&_goodHorizPosRelEstimateFact,     _goodHorizPosRelEstimateFactName);
-    _addFact(&_goodHorizPosAbsEstimateFact,     _goodHorizPosAbsEstimateFactName);
-    _addFact(&_goodVertPosAbsEstimateFact,      _goodVertPosAbsEstimateFactName);
-    _addFact(&_goodVertPosAGLEstimateFact,      _goodVertPosAGLEstimateFactName);
-    _addFact(&_goodConstPosModeEstimateFact,    _goodConstPosModeEstimateFactName);
-    _addFact(&_goodPredHorizPosRelEstimateFact, _goodPredHorizPosRelEstimateFactName);
-    _addFact(&_goodPredHorizPosAbsEstimateFact, _goodPredHorizPosAbsEstimateFactName);
-    _addFact(&_gpsGlitchFact,                   _gpsGlitchFactName);
-    _addFact(&_accelErrorFact,                  _accelErrorFactName);
-    _addFact(&_velRatioFact,                    _velRatioFactName);
-    _addFact(&_horizPosRatioFact,               _horizPosRatioFactName);
-    _addFact(&_vertPosRatioFact,                _vertPosRatioFactName);
-    _addFact(&_magRatioFact,                    _magRatioFactName);
-    _addFact(&_haglRatioFact,                   _haglRatioFactName);
-    _addFact(&_tasRatioFact,                    _tasRatioFactName);
-    _addFact(&_horizPosAccuracyFact,            _horizPosAccuracyFactName);
-    _addFact(&_vertPosAccuracyFact,             _vertPosAccuracyFactName);
 }
